:sparkles: feat: better add service feature using a tile
