+++ src/locales/en.json
@@ -357,7 +357,7 @@
     },
     "empty": "No connectors configured",
     "noAccount": "No account",
-    "voting": "You don't find the connector you were looking for? Please tell us!",
+    "add": "Add a service",
     "no-connectors-connected": "You have not yet connected connectors",
     "get-info": "Get automatically all your informations in your Cozy",
     "connect-account": "Connect my accounts",
