+++ src/styles/app.styl
@@ -40,10 +40,6 @@
     @extend $button
     @extend $button--regular
 
-.col-icon--add
-    width           16px
-    height          16px
-
 
 .col-account-connection-header
     @media (max-width: rem(930))
